refactor bank demo proof pack router
