+++ src/app/routers/bank_demo_proof.py
@@ -1,26 +1,13 @@
-from typing import Any
-
-from fastapi import APIRouter, Body, Header, status
+from fastapi import APIRouter, Header
 
 from app.contracts.bank_demo_proof import BankDemoProofEnvelopeResponse
-from app.middleware.correlation import correlation_id_var
+from app.routers.bank_demo_proof_common import (
+    BANK_DEMO_PROOF_RESPONSES,
+    bank_demo_correlation_id,
+)
 from app.services.advisory_service_provider import bank_demo_proof_service
 
 router = APIRouter(prefix="/api/v1/advisory/bank-demo-proof", tags=["bank-demo-proof"])
-
-BANK_DEMO_PROOF_RESPONSES: dict[int | str, dict[str, Any]] = {
-    status.HTTP_409_CONFLICT: {
-        "description": (
-            "lotus-advise rejected proof capture because material evidence does not match the "
-            "canonical RFC-0028 scenario."
-        )
-    },
-    422: {"description": "lotus-advise rejected the proof contract request shape."},
-}
-
-
-def _correlation_id(x_correlation_id: str | None) -> str:
-    return x_correlation_id or correlation_id_var.get() or ""
 
 
 @router.get(
@@ -38,7 +25,7 @@ async def get_bank_demo_scenario_contract(
     x_correlation_id: str | None = Header(default=None, alias="X-Correlation-Id"),
 ) -> BankDemoProofEnvelopeResponse:
     return await bank_demo_proof_service().get_scenario_contract(
-        correlation_id=_correlation_id(x_correlation_id),
+        correlation_id=bank_demo_correlation_id(x_correlation_id),
     )
 
 
@@ -58,32 +45,5 @@ async def get_bank_demo_supported_claim_register(
     x_correlation_id: str | None = Header(default=None, alias="X-Correlation-Id"),
 ) -> BankDemoProofEnvelopeResponse:
     return await bank_demo_proof_service().get_supported_claim_register(
-        correlation_id=_correlation_id(x_correlation_id),
-    )
-
-
-@router.post(
-    "/proof-packs",
-    response_model=BankDemoProofEnvelopeResponse,
-    summary="Build RFC-0028 Backend Proof Pack",
-    description=(
-        "Forwards RFC-0028 proof-pack capture requests to lotus-advise and returns the sanitized "
-        "proof bundle in a Gateway envelope. Advise remains proof authority; Gateway only "
-        "propagates correlation, problem status, and source-owned proof content."
-    ),
-    responses=BANK_DEMO_PROOF_RESPONSES,
-)
-async def build_bank_demo_proof_pack(
-    body: dict[str, Any] = Body(
-        ...,
-        description=(
-            "lotus-advise RFC-0028 proof-pack capture request containing governed live runtime "
-            "evidence and sanitized runtime posture."
-        ),
-    ),
-    x_correlation_id: str | None = Header(default=None, alias="X-Correlation-Id"),
-) -> BankDemoProofEnvelopeResponse:
-    return await bank_demo_proof_service().build_proof_pack(
-        body=body,
-        correlation_id=_correlation_id(x_correlation_id),
+        correlation_id=bank_demo_correlation_id(x_correlation_id),
     )
